# swagger.json
{
  "swagger": "2.0",
  "paths": {
...
    "/users/{id}": {
      "get": {
        "summary": "/users/{id}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {},
        "tags": [
          "users"
        ]
      },
      "delete": {
        "summary": "/users/{id}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {},
        "tags": [
          "users"
        ]
      },
      "put": {
        "summary": "/users/{id}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {},
        "tags": [
          "users"
        ]
      }
    },
...
  },
...
  "info": {
    "title": "hline",
    "version": "0.0.0",
    "license": {
      "name": "ISC"
    },
    "description": "Specification JSONs: [v2](/api-spec/v2), [v3](/api-spec/v3).\n\nTypeScript + Mongoose + MongoDB + Express API Server"
  }
}